copy utlitlies made working, home complete
