+++ src/lib/components/basic/icons/downArrow.svelte
@@ -8,6 +8,7 @@
       common: "#6E7681",
       indicator: "#74C7EC",
       profile: "#85889C",
+      profileActive: "#F2F2F0",
     };
     return typeToClassMap[type] || "";
   }
