+++ src/lib/components/basic/icons/profile.svelte
@@ -1,3 +1,7 @@
+<script>
+  export let color = "#85889C";
+</script>
+
 <svg
   width="30"
   height="30"
@@ -9,6 +13,6 @@
     fill-rule="evenodd"
     clip-rule="evenodd"
     d="M10 20C10 14.485 14.485 10 20 10C25.515 10 30 14.485 30 20C30 25.515 25.515 30 20 30C14.485 30 10 25.515 10 20ZM15.63 26.665L15.49 27.19H15.495C16.805 28.015 18.35 28.5 20.005 28.5C21.66 28.5 23.205 28.015 24.515 27.19L24.375 26.665C24.11 25.68 23.215 24.995 22.2 24.995H17.805C16.785 24.995 15.895 25.685 15.63 26.665ZM22.2 23.495C23.865 23.495 25.33 24.605 25.795 26.2V26.205C27.455 24.65 28.5 22.45 28.5 20C28.5 15.315 24.685 11.5 20 11.5C15.315 11.5 11.5 15.315 11.505 19.995C11.505 22.445 12.55 24.65 14.21 26.2C14.675 24.605 16.14 23.495 17.805 23.495H22.2ZM15.5 17.5C15.5 15.02 17.52 13 20 13C22.48 13 24.5 15.02 24.5 17.5C24.5 19.98 22.48 22 20 22C17.52 22 15.5 19.98 15.5 17.5ZM17 17.5C17 19.155 18.345 20.5 20 20.5C21.655 20.5 23 19.155 23 17.5C23 15.845 21.655 14.5 20 14.5C18.345 14.5 17 15.845 17 17.5Z"
-    fill="#85889C"
+    fill={color}
   />
 </svg>
